requirements.txt
streamlit
streamsync[ds]
pandas
matplotlib
seaborn
plotly
numpy
biopython
